requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
flask-login==0.6.2
reportlab==4.2.5
Pillow==10.1.0
python-dotenv==1.0.0
email-validator==2.1.0
gunicorn==21.2.0
